% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicforest.R
\name{classicforest}
\alias{classicforest}
\title{classic forest}
\usage{
classicforest(
  plotdata,
  madata,
  nsim = NULL,
  study_labels = NULL,
  summary_label = NULL,
  annotate_CI = FALSE,
  shaded_interval = NULL,
  statistic = NULL,
  confidence_level = 0.95,
  facet_titles = NULL,
  vline_intercept = 0,
  col = "Blues",
  summary_col = "Blues",
  tick_col = "firebrick",
  text_size = 3,
  x_lab = "Effect",
  y_lab = NULL,
  x_limit = NULL,
  x_breaks = NULL,
  jitter_nsim
)
}
\arguments{
\item{plotdata}{dataframe constructed by forest_constructor}

\item{madata}{dataframe constructed by forest_constructor}

\item{nsim}{(optional) column name that corresponds to simulation or bootstrap column. If specified, additional CI's of the individual statistics will be drawn.
If \code{nsim} is specified and no caption is set, a default caption will be set. Set \code{caption} to "" to override this functionality.}

\item{study_labels}{(optional) column name that corresponds to y-axis tick labels. If not specified, y-axis will be numbered.}

\item{annotate_CI}{logical. Default \code{TRUE}. Show a table next to the graph with the numeric values for the confidence interval.}

\item{shaded_interval}{numeric vector. Specified as c(lo, hi) for the interval you want to shade over. Default is NULL.}

\item{statistic}{is the actual statistic to output (i.e. median/mean)}

\item{vline_intercept}{(optional) numeric. Default 0.}

\item{text_size}{numeric. Text size for labels. Must be at least 3.5}

\item{x_lab}{string. x-axis label.}

\item{y_lab}{string. Default is not to label y axis.}

\item{x_limit}{numeric vector (c(lo, hi)) specifying the minimum and maximum values to show on the x-axis. See coord_cartesian() for details}

\item{x_breaks}{numeric vector of breaks to be used for the x-axis. See scale_x_continuous() for details.}

\item{jitter_nsim}{logical. Whether or not to vertically "jitter" the additional confidence intervals when using the \code{nsim} argument}
}
\description{
Private helper to generate the forest plot. Called internally by forest_constructor
}
\keyword{internal}
