% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest-constructor.R
\name{forest_constructor}
\alias{forest_constructor}
\title{Forest Constructor}
\usage{
forest_constructor(
  data,
  args,
  covariate = NULL,
  cov_level = NULL,
  nsim = NULL,
  summary_label = NULL,
  vline_intercept = 0,
  annotate_CI = TRUE,
  confidence_level,
  shaded_interval = NULL,
  statistic = NULL,
  x_lab = "Effect",
  y_lab = NULL,
  plot_width,
  facet_titles = NULL,
  CI_label = NULL,
  table_layout = NULL,
  text_size = 3,
  x_limit = NULL,
  x_breaks = NULL,
  jitter_nsim
)
}
\arguments{
\item{data}{a dataframe or tibble that contains the mean values and the lows & highs of the confidence interval}

\item{args}{mean/median, low CI, upper CI relating to the selected reference column}

\item{covariate}{(optional) column that defines subgroups within the data- all subgroups will be shown on the same plot but grouped together.}

\item{cov_level}{(optional) column name that corresponds to y-axis tick labels. If not specified, y-axis will be numbered.}

\item{nsim}{(optional) column name that corresponds to simulation or bootstrap column. If specified, additional CI's of the individual statistics will be drawn.
If \code{nsim} is specified and no caption is set, a default caption will be set. Set \code{caption} to "" to override this functionality.}

\item{vline_intercept}{(optional) numeric. Default 0.}

\item{annotate_CI}{logical. Default \code{TRUE}. Show a table next to the graph with the numeric values for the confidence interval.}

\item{shaded_interval}{numeric vector. Specified as c(lo, hi) for the interval you want to shade over. Default is NULL.}

\item{statistic}{is the actual statistic to output (i.e. median/mean)}

\item{x_lab}{string. x-axis label.}

\item{y_lab}{string. Default is not to label y axis.}

\item{plot_width}{numeric. Value between 1 and 12 to denote the ratio of plot : CI table}

\item{CI_label}{string. Ignored if \code{annotate_CI} is \code{FALSE}.}

\item{text_size}{numeric. Text size for labels. Must be at least 3.5}

\item{x_limit}{numeric vector (c(lo, hi)) specifying the minimum and maximum values to show on the x-axis. See coord_cartesian() for details}

\item{x_breaks}{numeric vector of breaks to be used for the x-axis. See scale_x_continuous() for details.}

\item{jitter_nsim}{logical. Whether or not to vertically "jitter" the additional confidence intervals when using the \code{nsim} argument}
}
\description{
Private helper to manipulate data and organize the structure of the forest plot and CI table
}
\keyword{internal}
