% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format-table-spacing.R
\name{format_table_spacing}
\alias{format_table_spacing}
\title{format table spacing}
\usage{
format_table_spacing(lab, text_size, title_fmt_val)
}
\arguments{
\item{lab}{dataframe constructed by table_plot}

\item{text_size}{numeric. Text size for labels. Must be at least 3.5}

\item{title_fmt_val}{additional space based on number of rows of CI_label}
}
\description{
Private helper to format vertical alignment of table. Called internally by table_plot
}
\keyword{internal}
