% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-forest.R
\name{plot_forest}
\alias{plot_forest}
\title{Function to create forest plots}
\usage{
plot_forest(
  data,
  stat = NULL,
  statistic = c("median", "mean"),
  CI = 0.95,
  covariate = NULL,
  cov_level = NULL,
  metagroup = NULL,
  nsim = NULL,
  summary_label = NULL,
  vline_intercept = 0,
  annotate_CI = TRUE,
  shaded_interval = NULL,
  sigfig = 2,
  x_lab = "Effect",
  y_lab = NULL,
  CI_label = NULL,
  text_size = 3.5,
  plot_width = 8,
  caption = NULL,
  x_breaks = NULL,
  x_limit = NULL,
  jitter_nsim = FALSE,
  ...
)
}
\arguments{
\item{data}{a dataframe or tibble that contains the mean values and the lows & highs of the confidence interval}

\item{stat}{is the column to perform calculations on (i.e. median/mean, lower, and upper CI)}

\item{statistic}{is the actual statistic to output (i.e. median/mean)}

\item{CI}{is the confidence interval to plot}

\item{covariate}{(optional) column that defines subgroups within the data- all subgroups will be shown on the same plot but grouped together.}

\item{cov_level}{(optional) column name that corresponds to y-axis tick labels. If not specified, y-axis will be numbered.}

\item{metagroup}{(optional) column name that corresponds to metagroups. Similar to facet wrap. Will produce independent plots per metagroup.}

\item{nsim}{(optional) column name that corresponds to simulation or bootstrap column. If specified, additional CI's of the individual statistics will be drawn.
If \code{nsim} is specified and no caption is set, a default caption will be set. Set \code{caption} to "" to override this functionality.}

\item{vline_intercept}{(optional) numeric. Default 0.}

\item{annotate_CI}{logical. Default \code{TRUE}. Show a table next to the graph with the numeric values for the confidence interval.}

\item{shaded_interval}{numeric vector. Specified as c(lo, hi) for the interval you want to shade over. Default is NULL.}

\item{sigfig}{numeric. Number of significant digits to round the table values to.}

\item{x_lab}{string. x-axis label.}

\item{y_lab}{string. Default is not to label y axis.}

\item{CI_label}{string. Ignored if \code{annotate_CI} is \code{FALSE}.}

\item{text_size}{numeric. Text size for labels. Must be at least 3.5}

\item{plot_width}{numeric. Value between 1 and 12 to denote the ratio of plot : CI table}

\item{caption}{string. A patchwork styled caption for the overall plot.}

\item{x_breaks}{numeric vector of breaks to be used for the x-axis. See scale_x_continuous() for details.}

\item{x_limit}{numeric vector (c(lo, hi)) specifying the minimum and maximum values to show on the x-axis. See coord_cartesian() for details}

\item{jitter_nsim}{logical. Whether or not to vertically "jitter" the additional confidence intervals when using the \code{nsim} argument}

\item{...}{additional args passed to \code{patchwork::wrap_plots()} for metagrouped plots.}

\item{summary_labels}{(optional) labeler function created using \code{ggplot2::as_labeller}. Labels for group and metagroup.}
}
\description{
Creates a forest plot with or without confidence intervals
Expects..
See params below..
}
