% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-data.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize Data}
\usage{
summarize_data(
  data,
  value,
  group,
  group_level = NULL,
  metagroup = NULL,
  replicate = NULL,
  CI = 0.95,
  statistic = c("median", "mean")
)
}
\arguments{
\item{data}{A dataframe or tibble to summarize. See Details section for required format.}

\item{value}{name of the column in \code{data} to perform calculations on (i.e.
median/mean, lower, and upper CI)}

\item{group}{name of the column in \code{data} that defines groups within the
data. Often, this will contain the names of the covariates you are grouping
by.}

\item{group_level}{(optional) name of the column in \code{data} that contains
subgroups to group by. For example, if your \code{group} column contains
covariates like \code{WEIGHT} and \code{AGE}, this column could contain categories
like \code{underweight}, \code{average}, \code{overweight}, \code{young}, \code{mid}, \code{elderly}, etc.}

\item{metagroup}{(optional) name of the column in \code{data} that contains
\code{metagroups}. Similar to facet wrap, if passed, this will cause
\code{\link[=plot_forest]{plot_forest()}} to produce independent plots per metagroup.}

\item{replicate}{(optional) name of the column in \code{data} that contains to an
index of replicates, for example with multiple simulations or
bootstrapping. If specified, \code{\link[=plot_forest]{plot_forest()}} will draw additional CI's of
the individual statistics, as small lines above each primary line.}

\item{CI}{numeric between 0 and 1. This is the confidence interval to plot, defaults to \code{0.95}.}

\item{statistic}{is the actual statistic to output (i.e. median/mean)}
}
\description{
Summarize input data to prepare for passing to \code{\link[=plot_forest]{plot_forest()}}. Takes a
data.frame or tibble, calculates the relevant confidence intervals, and
returns a tibble that can be passed directly to \code{\link[=plot_forest]{plot_forest()}}. See Details
section for data specification and format.
}
\details{
\strong{Input Data}

The tibble passed to \code{data} must be in a "long" format and will have 2-5
columns: \code{value}, \code{group}, and optionally any of \code{group_level}, \code{metagroup},
and/or \code{replicate}. These are each described in detail in the input arguments
section.

\strong{Output Data}
The tibble output from this function will have one of two formats, depending on whether
\code{replicate} was passed (details below).

Either way, the output tibble will have a column named \code{group}, containing the
values in the column you passed to the \code{group} argument, and optionally similar columns
for \code{group_level} and \code{metagroup} if those were passed.
\itemize{
\item \strong{Without \code{replicate}} If \code{replicate} is \emph{not} passed, the output data
}
}
