% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-plot-new-wip.R
\name{table_plot2}
\alias{table_plot2}
\title{table plot}
\usage{
table_plot2(
  plotdata,
  r = 5.5,
  l = 5.5,
  tbl_titles = NULL,
  text_size,
  y_limit,
  y_breaks,
  y_lines
)
}
\arguments{
\item{plotdata}{dataframe constructed by forest_constructor}

\item{r, l}{parameters corresponding to the plot margin for the confidence intervals}

\item{text_size}{numeric. Text size for labels. Must be at least 3.5}

\item{madata}{dataframe constructed by forest_constructor}
}
\description{
Private helper to record and tabulate confidence intervals. Called internally by forest_constructor
}
\keyword{internal}
