% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classicforest.R
\name{classicforest}
\alias{classicforest}
\title{classic forest}
\usage{
classicforest(
  plotdata,
  madata,
  nsim = NULL,
  group_level = NULL,
  summary_label = NULL,
  annotate_CI = FALSE,
  shaded_interval = NULL,
  facet_titles = NULL,
  vline_intercept = 0,
  col = "Blues",
  summary_col = "Blues",
  tick_col = "firebrick",
  text_size = 3,
  x_lab = "Effect",
  y_lab = NULL,
  x_limit = NULL,
  x_breaks = NULL,
  jitter_reps,
  shapes,
  ggplot_theme
)
}
\arguments{
\item{plotdata}{dataframe constructed by forest_constructor}

\item{madata}{dataframe constructed by forest_constructor}

\item{summary_label}{(optional) labeler function created using \code{ggplot2::as_labeller}. Labels for group and metagroup.}

\item{annotate_CI}{logical. Default \code{TRUE}. Show a table next to the graph with the numeric values for the confidence interval.}

\item{shaded_interval}{numeric vector. Specified as c(lo, hi) for the interval you want to shade over. Default is NULL.}

\item{vline_intercept}{(optional) numeric. Default 0.}

\item{text_size}{numeric. Text size for labels. Must be at least 3.5}

\item{x_lab}{string. x-axis label.}

\item{y_lab}{string. Default is not to label y axis.}

\item{x_limit}{numeric vector (c(lo, hi)) specifying the minimum and maximum values to show on the x-axis. See coord_cartesian() for details}

\item{x_breaks}{numeric vector of breaks to be used for the x-axis. See scale_x_continuous() for details.}

\item{jitter_reps}{logical. Whether or not to vertically "jitter" the
additional confidence intervals when using multiple replicates (i.e. when
input data has 9 numeric columns instead of 3).}

\item{shapes}{string. Denotes the shape of the mean/median value, passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.
Can be one of the following: "square", "diamond", "circle", or "triangle".}

\item{ggplot_theme}{a ggplot theme. Note that some legend options, text size, and grid elements may be overwritten.
See \code{\link[ggplot2:theme]{ggplot2::theme()}} for details.}
}
\description{
Private helper to generate the forest plot. Called internally by forest_constructor
}
\keyword{internal}
