% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest-constructor.R
\name{forest_constructor}
\alias{forest_constructor}
\title{Forest Constructor}
\usage{
forest_constructor(
  data,
  args,
  nsim = NULL,
  summary_label = NULL,
  vline_intercept = 0,
  annotate_CI = TRUE,
  sigfig,
  shaded_interval = NULL,
  x_lab = "Effect",
  y_lab = NULL,
  plot_width,
  facet_titles = NULL,
  CI_label = NULL,
  table_layout = NULL,
  text_size = 3,
  x_limit = NULL,
  x_breaks = NULL,
  jitter_reps
)
}
\arguments{
\item{data}{a dataframe or tibble that contains the summarized data you want
to plot. This must be in the same format as the tibble that is output by
\code{\link[=summarize_data]{summarize_data()}}. See "Output Data" in Details section of \code{\link[=summarize_data]{summarize_data()}} documentation.}

\item{args}{mean/median, low CI, upper CI relating to the selected reference column}

\item{vline_intercept}{(optional) numeric. Default 0.}

\item{annotate_CI}{logical. Default \code{TRUE}. Show a table next to the graph with the numeric values for the confidence interval.}

\item{sigfig}{numeric. Number of significant digits to round the table values to.}

\item{shaded_interval}{numeric vector. Specified as c(lo, hi) for the interval you want to shade over. Default is NULL.}

\item{x_lab}{string. x-axis label.}

\item{y_lab}{string. Default is not to label y axis.}

\item{plot_width}{numeric. Value between 1 and 12 to denote the ratio of plot : CI table}

\item{CI_label}{string. Ignored if \code{annotate_CI} is \code{FALSE}.}

\item{text_size}{numeric. Text size for labels. Must be at least 3.5}

\item{x_limit}{numeric vector (c(lo, hi)) specifying the minimum and maximum values to show on the x-axis. See coord_cartesian() for details}

\item{x_breaks}{numeric vector of breaks to be used for the x-axis. See scale_x_continuous() for details.}

\item{jitter_reps}{logical. Whether or not to vertically "jitter" the
additional confidence intervals when using multiple replicates (i.e. when
input data has 9 numeric columns instead of 3).}
}
\description{
Private helper to manipulate data and organize the structure of the forest plot and CI table
}
\keyword{internal}
