% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-forest.R
\name{plot_forest}
\alias{plot_forest}
\title{Create forest plots}
\usage{
plot_forest(
  data,
  summary_label = NULL,
  vline_intercept = 0,
  annotate_CI = TRUE,
  shaded_interval = NULL,
  digits = 3,
  maxex = NULL,
  x_lab = "Effect",
  y_lab = NULL,
  CI_label = NULL,
  caption = NULL,
  text_size = 3.5,
  plot_width = 8,
  x_breaks = NULL,
  x_limit = NULL,
  jitter_reps = FALSE,
  shapes = c("square", "diamond", "circle", "triangle"),
  ggplot_theme = ggplot2::theme_bw(),
  ...
)
}
\arguments{
\item{data}{a dataframe or tibble that contains the summarized data you want
to plot. This must be in the same format as the tibble that is output by
\code{\link[=summarize_data]{summarize_data()}}. See "Output Data" in Details section of \code{\link[=summarize_data]{summarize_data()}} documentation.}

\item{summary_label}{(optional) labeler function created using \code{ggplot2::as_labeller}. Labels for group and metagroup.}

\item{vline_intercept}{(optional) numeric. Default 0.}

\item{annotate_CI}{logical. Default \code{TRUE}. Show a table next to the graph with the numeric values for the confidence interval.}

\item{shaded_interval}{numeric vector. Specified as c(lo, hi) for the interval you want to shade over. Default is NULL.}

\item{digits}{numeric. Number of significant digits to round the table values to. Passed through to \code{\link[pmtables:sig]{pmtables::sig()}}}

\item{maxex}{numeric. Maximum number of significant digits before moving to scientific notation. Passed through to \code{\link[pmtables:sig]{pmtables::sig()}}.}

\item{x_lab}{string. x-axis label.}

\item{y_lab}{string. Default is not to label y axis.}

\item{CI_label}{string. Ignored if \code{annotate_CI} is \code{FALSE}.}

\item{caption}{string. A patchwork styled caption for the overall plot.}

\item{text_size}{numeric. Text size for labels. Must be at least 3.5}

\item{plot_width}{numeric. Value between 1 and 12 to denote the ratio of plot : CI table}

\item{x_breaks}{numeric vector of breaks to be used for the x-axis. See scale_x_continuous() for details.}

\item{x_limit}{numeric vector (c(lo, hi)) specifying the minimum and maximum values to show on the x-axis. See coord_cartesian() for details}

\item{jitter_reps}{logical. Whether or not to vertically "jitter" the
additional confidence intervals when using multiple replicates (i.e. when
input data has 9 numeric columns instead of 3).}

\item{shapes}{string. Denotes the shape of the mean/median value, passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.
Can be one of the following: "square", "diamond", "circle", or "triangle".}

\item{ggplot_theme}{a ggplot theme. Note that some legend options, text size, and grid elements may be overwritten.
See \code{\link[ggplot2:theme]{ggplot2::theme()}} for details.}

\item{...}{additional args passed to \code{patchwork::wrap_plots()} for metagrouped plots.}
}
\description{
Creates a forest plot with or without confidence intervals.
See \href{https://metrumresearchgroup.github.io/pmforest/articles/getting-started.html#summarize-data}{the "Getting Started" vignette}
for usage and examples.
}
