% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-data.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize Data}
\usage{
summarize_data(
  data,
  stat,
  covariate,
  cov_level,
  metagroup,
  nsim,
  CI = CI,
  statistic
)
}
\arguments{
\item{data}{a dataframe or tibble that contains the mean values and the lows & highs of the confidence interval}

\item{stat}{is the column to perform calculations on (i.e. median/mean, lower, and upper CI)}

\item{covariate}{(optional) column that defines subgroups within the data- all subgroups will be shown on the same plot but grouped together.}

\item{cov_level}{(optional) column name that corresponds to y-axis tick labels. If not specified, y-axis will be numbered.}

\item{metagroup}{(optional) column name that corresponds to metagroups. Similar to facet wrap. Will produce independent plots per metagroup.}

\item{nsim}{(optional) column name that corresponds to simulation or bootstrap column. If specified, additional CI's of the individual statistics will be drawn.
If \code{nsim} is specified and no caption is set, a default caption will be set. Set \code{caption} to "" to override this functionality.}

\item{CI}{is the confidence interval to plot}

\item{statistic}{is the actual statistic to output (i.e. median/mean)}

\item{r, l}{parameters corresponding to the plot margin for the confidence intervals}
}
\description{
Private helper to summarize the referenced column. Called internally by plot_forest
}
\keyword{internal}
