% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize-data.R
\name{summarize_data}
\alias{summarize_data}
\title{Summarize Data}
\usage{
summarize_data(
  data,
  value,
  group,
  group_level = NULL,
  metagroup = NULL,
  replicate = NULL,
  probs = c(0.05, 0.95),
  statistic = c("median", "mean", "geo_mean"),
  rep_probs = c(0.025, 0.975),
  rep_statistic = c("median", "mean", "geo_mean")
)
}
\arguments{
\item{data}{A dataframe or tibble to summarize. See Details section for required format.}

\item{value}{name of the column in \code{data} to perform calculations on (i.e.
median/mean, lower, and upper CI)}

\item{group}{name of the column in \code{data} that defines groups within the
data. Often, this will contain the names of the covariates you are grouping
by.}

\item{group_level}{(optional) name of the column in \code{data} that contains
subgroups to group by. For example, if your \code{group} column contains
covariates like \code{WEIGHT} and \code{AGE}, this column could contain categories
like \code{underweight}, \code{average}, \code{overweight}, \code{young}, \code{mid}, \code{elderly}, etc.}

\item{metagroup}{(optional) name of the column in \code{data} that contains
\code{metagroups}. Similar to facet wrap, if passed, this will cause
\code{\link[=plot_forest]{plot_forest()}} to produce independent plots per metagroup.}

\item{replicate}{(optional) name of the column in \code{data} that contains to an
index of replicates, for example with multiple simulations or
bootstrapping. If specified, \code{\link[=plot_forest]{plot_forest()}} will draw additional CI's of
the individual statistics, as small lines above each primary line.}

\item{probs}{numeric vector of length two, both between 0 and 1, corresponding to
your lower and upper tail probabilities. Defaults to \code{c(0.05, 0.95)}}

\item{statistic}{is the actual statistic to output (i.e. median/mean)}

\item{rep_probs}{same as \code{probs} but used \strong{only when \code{replicate} is passed}
for the minor intervals (i.e. the small lines) above the major interval (i.e. the big lines).}

\item{rep_statistic}{same as \code{statistic} but used \strong{only when \code{replicate} is
passed} for the minor intervals (i.e. the small lines) above the major interval (i.e. the big lines).}
}
\description{
Summarize input data to prepare for passing to \code{\link[=plot_forest]{plot_forest()}}. Takes a
data.frame or tibble, calculates the relevant confidence intervals, and
returns a tibble that can be passed directly to \code{\link[=plot_forest]{plot_forest()}}. See Details
section for data specification and format.
}
\details{
\strong{Input Data}

The tibble passed to \code{data} must be in a "long" format and has 2-5
columns: \code{value}, \code{group}, and optionally any of \code{group_level}, \code{metagroup},
and/or \code{replicate}. These are each described in detail in the input arguments
section.

\strong{Output Data}

The tibble output from this function has one of two formats, depending on whether
\code{replicate} was passed (details below).

Either way, the output tibble has a column named \code{group}, containing the
values in the column you passed to the \code{group} argument, and optionally analogous columns
for \code{group_level} and \code{metagroup} if those were passed.
\itemize{
\item \strong{Without \code{replicate}} If \code{replicate} is \emph{not} passed, the output data has three
additional columns \code{mid}, \code{lo}, and \code{hi}, containing the summarized values corresponding
to what was passed to \code{statistic} (\code{mid}) and \code{probs} (\code{lo}/\code{hi}).
\item \strong{With \code{replicate}} If \code{replicate} \emph{is} passed, the output data has
nine additional columns \code{mid_mid}, \code{mid_lo}, \code{mid_hi}, plus three more each
for \verb{lo_*} and \verb{hi_*}, containing the summarized values. In this case, the
\code{mid_mid}, \code{lo_mid}, and \code{hi_mid} correspond to the values of the major
interval (i.e. the big lines and data point) and the \verb{*_mid}, \verb{*_lo}, and
\verb{*_hi} correspond to the values for each minor interval (i.e. the small
lines).
}
}
