% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-plot.R
\name{table_plot}
\alias{table_plot}
\title{table plot}
\usage{
table_plot(
  tbl,
  summary_label,
  ID,
  r = 5.5,
  l = 5.5,
  tbl_titles = NULL,
  plotdata,
  text_size,
  y_limit,
  y_breaks
)
}
\arguments{
\item{summary_label}{(optional) labeler function created using \code{ggplot2::as_labeller}. Labels for group and metagroup.}

\item{r, l}{parameters corresponding to the plot margin for the confidence intervals}

\item{plotdata}{dataframe constructed by forest_constructor}

\item{text_size}{numeric. Text size for labels. Must be at least 3.5}

\item{madata}{dataframe constructed by forest_constructor}
}
\description{
Private helper to record and tabulate confidence intervals. Called internally by forest_constructor
}
\keyword{internal}
